Move locus and orthogroup views to their own module.
